Split out cell renderer
