+++ packages/data-sheet/src/index.ts
@@ -4,3 +4,5 @@ export * from "./components";
 export * from "./provider";
 export * from "./utils";
 export * from "./types";
+
+export { Cell } from "@blueprintjs/table";
